[sin sprint] ya se borran los informes cuando se tienen que borrar, falta corregir error de que cuando se crea un informe da un fallo, aunke aun así se crea!, y darle estilos

+++ src/Edukagames/UserBundle/Util/SaveEraseFile.php
@@ -23,8 +23,9 @@ class SaveEraseFile
     
     public static function eraseFile($dir)
     {
-    	if (file_exists($dir))
-    		unlink($dir);
+    	if ( !is_dir($dir))
+    		if (file_exists($dir))
+    			unlink($dir);
     }
 
 }
